requirements.txt
matplotlib==3.5.1
numpy==1.21.5
patchify==0.2.3
Pillow==10.3.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
